move enum derivation to vulcan-generic
