+++ modules/core/src/test/scala-2/vulcan/examples/CaseClassAvroDoc.scala
@@ -1,11 +1,13 @@
 package vulcan.examples
 
 import vulcan.{AvroDoc, Codec}
+import scala.annotation.nowarn
 
 @AvroDoc("documentation")
 final case class CaseClassAvroDoc(value: Option[String])
 
 object CaseClassAvroDoc {
+  @nowarn("msg=deprecated")
   implicit val codec: Codec[CaseClassAvroDoc] =
     Codec.deriveEnum(
       symbols = List("first"),
