+++ modules/core/src/test/scala-2/vulcan/examples/FixedNamespace.scala
@@ -1,5 +1,6 @@
 package vulcan.examples
 
+import scala.annotation.nowarn
 import cats.Eq
 import org.scalacheck.{Arbitrary, Gen}
 import vulcan.{AvroDoc, AvroNamespace, Codec}
@@ -9,6 +10,7 @@ import vulcan.{AvroDoc, AvroNamespace, Codec}
 final case class FixedNamespace(bytes: Array[Byte])
 
 object FixedNamespace {
+  @nowarn("msg=deprecated")
   implicit val codec: Codec[FixedNamespace] =
     Codec.deriveFixed(
       size = 1,
