--- modules/generic/src/test/scala/vulcan/generic/EitherValues.scala
@@ -1,7 +1,0 @@
-package vulcan.generic
-
-trait EitherValues {
-  implicit final class EitherValuesSyntax[A, B](val e: Either[A, B]) {
-    def value: B = e.getOrElse(throw new NoSuchElementException(s"Expected Right was $e"))
-  }
-}
