+++ modules/generic/src/test/scala/vulcan/generic/examples/CaseClassFieldAvroDoc.scala
@@ -1,4 +1,4 @@
-package vulcan.examples
+package vulcan.generic.examples
 
 import vulcan.Codec
 import vulcan.generic._
